NEW: Start mapping point geometry
