+++ coresdk/src/coresdk/geometry.cpp
@@ -72,14 +72,6 @@ vector_2d vector_to(float x, float y)
     return result;
 }
 
-point_2d point_at(float x, float y)
-{
-    point_2d result;
-    result.x = x;
-    result.y = y;
-    return result;
-}
-
 circle circle_at(const point_2d &pt, float radius)
 {
     circle result;
@@ -123,23 +115,3 @@ quad quad_from(point_2d p1, point_2d p2, point_2d p3, point_2d p4)
     return result;
 }
 
-point_2d random_screen_point()
-{
-    return random_window_point(current_window());
-}
-
-point_2d random_window_point(window wind)
-{
-    point_2d result;
-    result.x = rnd() * screen_width();
-    result.y = rnd() * screen_height();
-    return result;
-}
-
-point_2d random_bitmap_point(bitmap bmp)
-{
-    point_2d result;
-    result.x = rnd() * bitmap_width(bmp);
-    result.y = rnd() * bitmap_height(bmp);
-    return result;
-}
